fixed: folder not found on deletion process. added 'ab' to custom stopwords.

+++ DiskStorageMisc.py
@@ -29,15 +29,16 @@ class DiskStorageMisc:
     # deletes all files from folder
     @staticmethod
     def delete_from_folder(path):
-        for filename in os.listdir(path):
-            file_path = os.path.join(path, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                SessionLogger.log('Failed to delete %s. Reason: %s' % (file_path, e))
+        if os.path.exists(path):
+            for filename in os.listdir(path):
+                file_path = os.path.join(path, filename)
+                try:
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.unlink(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                except Exception as e:
+                    SessionLogger.log('Failed to delete %s. Reason: %s' % (file_path, e))
 
     # expects session identifier
     # deletes all data from session
